changed some comments in execution to the .h file
