--- o.txt
@@ -1,30 +1,0 @@
-Makefile
-README.md
-TODO
-docs
-en.subject.pdf
-evaluation_minishell.pdf
-includes
-libs
-minishell
-o.txt
-obj
-src
-tester.py
-testprogram.c
-tests
-Makefile
-README.md
-TODO
-docs
-en.subject.pdf
-evaluation_minishell.pdf
-includes
-libs
-minishell
-o.txt
-obj
-src
-tester.py
-testprogram.c
-tests
