--- out.txt
@@ -1,16 +1,0 @@
-Makefile
-README.md
-TODO
-docs
-en.subject.pdf
-evaluation_minishell.pdf
-includes
-libs
-minishell
-o.txt
-obj
-out.txt
-src
-tester.py
-testprogram.c
-tests
