+++ src/core/shell.c
@@ -6,7 +6,7 @@
 /*   By: asioud <asioud@42heilbronn.de>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/02 01:45:52 by asioud            #+#    #+#             */
-/*   Updated: 2023/06/18 00:58:45 by asioud           ###   ########.fr       */
+/*   Updated: 2023/06/18 01:14:41 by asioud           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -97,4 +97,3 @@ int parse_and_execute(t_cli *cli)
 	// free_node_tree(ast_cmd);
 	return (0);
 }
-
